third_party/blink/web_tests/external/wpt/svg/linking/scripted/a.hreflang-getter-01-expected.txt

This is a testharness.js-based test.
[FAIL] Test anchor's hreflang getter
  assert_equals: expected (string) "en" but got (undefined) undefined
Harness: the test ran to completion.
